Switch from "unreleased" to running version
Now, the first changelog entry is still unreleased changes, but it
already has a version number assigned. The benefit is a release
can be created at any time by creating an appropriate tag.
The downsides are that it's more difficult (slash impossible) to
determine what's released and what isn't from the changelog.